Separate settings code into its own files
